Created Pricing page
-Added an Services folder in Components and this will  house
an AllServices component that shows all pricing options
Osiris offers. I also put the more MoreServicesButton
that is used on the about page here.
-This component uses the Pricing.js from the layout folder with another
three pricing cards added to it. I will keep the Pricing.js for the
home page, and use this for when a user actually clicks on the pricing
tab or MoreServicesButton.

-I created a PricingPage in the pages folder and imported the
Navigation, AllServices, and Footer components. Still need to finish
styling this page.